Address mac compiler error

+++ src/module_library/leaf_energy_balance.h
@@ -17,7 +17,7 @@ struct energy_balance_outputs {
     double RH_canopy;        //!< Relative humidity in the canopy, just outside the leaf boundary layer (dimensionless)
     double storage;          //!< Rate of energy storage by the leaf; should be zero (J / m^2 / s)
     double TransR;           //!< Transpiration rate (mmol / m^2 / s)
-    int iterations;          //!< Number of iterations used by convergence loop
+    size_t iterations;       //!< Number of iterations used by convergence loop
 };
 
 energy_balance_outputs leaf_energy_balance(
